requirements.txt
flask
flask-cors
python-dotenv
faiss-cpu
sentence-transformers
transformers
torch
pymongo
flask-limiter
bleach
